+++ packages/components/label/src/Label.stories.tsx
@@ -10,13 +10,47 @@ const meta: Meta<typeof Label> = {
 export default meta
 
 export const Default: StoryFn = _args => (
-  <div className="gap-md flex flex-col">
-    <Label htmlFor="name">First name</Label>
+  <div className="flex flex-col gap-md">
+    <Label htmlFor="label-default">Title</Label>
     <input
       type="text"
-      id="name"
-      placeholder="Jon Snow"
-      className="p-md border-neutral active:border-primary border-md rounded-sm"
+      id="label-default"
+      placeholder="IPhone 14"
+      className="rounded-sm border-md border-neutral p-md active:border-primary"
+    />
+  </div>
+)
+
+export const Required: StoryFn = _args => (
+  <div className="flex flex-col gap-md">
+    <Label className="flex items-center gap-sm" htmlFor="label-required">
+      Title
+      <Label.RequiredIndicator />
+    </Label>
+
+    <input
+      type="text"
+      id="label-required"
+      placeholder="IPhone 14"
+      className="rounded-sm border-md border-neutral p-md active:border-primary"
+      required
+    />
+  </div>
+)
+
+export const RequiredIndicator: StoryFn = _args => (
+  <div className="flex flex-col gap-md">
+    <Label className="flex items-center gap-sm" htmlFor="label-indicator">
+      Title
+      <Label.RequiredIndicator>Required</Label.RequiredIndicator>
+    </Label>
+
+    <input
+      type="text"
+      id="label-indicator"
+      placeholder="IPhone 14"
+      className="rounded-sm border-md border-neutral p-md active:border-primary"
+      required
     />
   </div>
 )