+++ packages/components/button/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.12.1](https://github.com/adevinta/spark/compare/@spark-ui/button@1.12.0...@spark-ui/button@1.12.1) (2023-06-16)
+
+**Note:** Version bump only for package @spark-ui/button
+
 # [1.12.0](https://github.com/adevinta/spark/compare/@spark-ui/button@1.11.11...@spark-ui/button@1.12.0) (2023-06-14)
 
 ### Features